+++ web/packages/pyxis-user-site/src/store.ts
@@ -1,0 +1,25 @@
+import { configureStore } from '@reduxjs/toolkit';
+import { setupListeners } from '@reduxjs/toolkit/query';
+import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
+import { publicApi } from './api/publicApi';
+
+export function makeStore() {
+  return configureStore({
+    reducer: {
+      [publicApi.reducerPath]: publicApi.reducer,
+    },
+    middleware: (getDefaultMiddleware) =>
+      getDefaultMiddleware().concat(publicApi.middleware),
+  });
+}
+
+export const store = makeStore();
+
+setupListeners(store.dispatch);
+
+export type AppStore = ReturnType<typeof makeStore>;
+export type RootState = ReturnType<AppStore['getState']>;
+export type AppDispatch = AppStore['dispatch'];
+
+export const useAppDispatch = () => useDispatch<AppDispatch>();
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;